requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Frontend
streamlit==1.39.0

# Database drivers
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.2
neo4j==5.25.0

# AI APIs
anthropic==0.40.0
openai==1.54.0

# Document parsing
pymupdf==1.24.11
python-docx==1.1.2

# Config / validation
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP client (UI → API calls)
httpx==0.27.2

# Retry / resilience
tenacity==9.0.0

# PDF export
fpdf2==2.8.1

# Multipart uploads
python-multipart==0.0.12

# Async file I/O
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
